Day04

+++ Day04/spring-calculator/src/main/java/com/sankalp/calculator/CalculatorController.java
@@ -8,19 +8,19 @@ import org.springframework.web.bind.annotation.RestController;
 public class CalculatorController {
 
     @GetMapping("/add")
-    public float addition(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
+    public double addition(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
         return a+b;
     }
     @GetMapping("/sub")
-    public float subtraction(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
+    public double subtraction(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
         return a-b;
     }
     @GetMapping("/mul")
-    public float multiplication(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
+    public double multiplication(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
         return a*b;
     }
     @GetMapping("/div")
-    public float division(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
+    public double division(@RequestParam(name="a", required = true) double a,@RequestParam(name="b", required = true) double b){
         return a/b;
     }
 }
